+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q049/Main.java
@@ -1,0 +1,47 @@
+package com.noob.algorithm.leetcode.q049;
+
+
+import com.noob.algorithm.base.util.PrintUtil;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Main {
+
+    // 初始化测试用例
+    public static List<String[]> initCase() {
+        List<String[]> caseList = new ArrayList<>();
+
+        // todo 补充测试用例集合
+        caseList.add(new String[]{""});
+
+        return caseList;
+    }
+
+    public static void testSolution1(List<String[]> caseList) {
+
+        for (String[] strs : caseList) {
+            Solution1 solution = new Solution1();
+            List<List<String>> res = solution.groupAnagrams(strs);
+            // PrintUtil.printMatrix(res);
+        }
+
+    }
+
+
+    public static void testSolution2(List<String[]> caseList) {
+
+    }
+
+
+    public static void main(String[] args) {
+        List<String[]> caseList = initCase();
+
+        testSolution1(caseList);
+
+        System.out.println("-------------------------------");
+
+        testSolution2(caseList);
+
+    }
+}
